Bye.sh final commit (Project Completed.)
Haz some public ascii copy pasted from the depths of internet. I don't really know why, I am even pushing this to github.. Elementary level shell script for coping with the laziness of moving mouse and clicking the logout button. 

Not a full-proof script and it can be a tragic comedy if $USER returns root. For the sakes of it since I know the uses of this script I am never going to run it in sudo. I really want to leave the doors open for sudo Bye.sh exec. :D

aka, if [ $USER = root ] is probably a check you should do. But since I don't know - perhaps you want to nuke the hell out of your root; I have left it open. This script is supposed to be a simple one and it is not my job to make the adjustments. It's under GNU license = feel free to edit.